change temperature type in decider variables

+++ src/main/java/de/griesser/outfit/decider/api/Variables.java
@@ -2,12 +2,10 @@ package de.griesser.outfit.decider.api;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class Variables {
 
     // field names must match input expression texts from dmn decision table
-    private final BigDecimal temperature;
+    private final double temperature;
 
 }
